+++ CotEditor/Sources/PreferencesTabViewController.swift
@@ -126,7 +126,6 @@ private extension PreferencesTabViewController {
         "File Drop": "aO6-oS-ZGt",
         "Key Bindings": "b8q-WN-1ls",
         "Print": "UuB-iq-kOt",
-        "Integration": "gEv-qP-tRM",
     ]